+++ src/Utility/CodeLoader.h
@@ -29,7 +29,7 @@ public:
 
   // Store the contents of the given file into the component of the tile at
   // the given position.
-  static void loadCode(string& filename, Tile& tile, int position);
+  static void loadCode(string& filename, Tile& tile, uint position);
 
   // Store the contents of the file into the given component.
   static void loadCode(string& filename, WrappedTileComponent& component);
@@ -43,8 +43,8 @@ public:
 
   static vector<Word>& getData(string& filename);
 private:
-  static int fileType(string& filename);
-  static Word getWord(std::ifstream& file, int type);
+  static uint8_t fileType(string& filename);
+  static Word getWord(std::ifstream& file, uint8_t type);
 
   enum FILETYPE {LOKI, BINARY, DATA};
